requirements.txt
langchain
openai
chromadb
tiktoken
python-dotenv
fastapi
uvicorn
streamlit
langchain-openai
langchain-community
